Build: Build the site if not already built and pass next command to `site`
We must use `rebuild` instead of `build` in case there's an existing
`_cache` directory. (`build` doesn't rebuild in that case.)

After that we pass the argument given on the command line to Hakyll's
`site` command; if no argument is given it will print help. Typically
you'd pass `server` or `watch`.

This takes `-C` to clean everything